Fixed bug in code so that enemies do not spawn in greater number than the level total allows.
Updated projectile code to spawn and launch in correct direction of the Firing Socket of the player weapon.  Updated code to use OnComponentBeginOverlap not OnComponentHit as the latter had varying success.

Added new custom trace channel for projectiles.  Set default collision for enemies only to use this trace channel for overlapping.